+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/swdmplog/TxnSwdmplog.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TxnSwdmplog {
 	
-	private Integer idMovTxnSwdmplog;
+	private Long idMovTxnSwdmplog;
 	private String messageType;
 	private String stfwdIndicator;
 	private String pan;
